Fixed errors

+++ src/app/api/delete/[id]/route.ts
@@ -1,19 +1,28 @@
 import { prisma } from "@/lib/prisma";
 import { NextResponse } from "next/server";
 
-interface IDeleteParams {
-  params: {
-    id: string;
-  };
-}
+export const DELETE = async (req: Request) => {
+  try {
+    const url = new URL(req.url);
+    const id = url.pathname.split("/").pop();
 
-export const DELETE = async (req: Request, { params }: IDeleteParams) => {
-  const { id } = params;
+    if (!id) {
+      return NextResponse.json({ error: "ID is required" }, { status: 400 });
+    }
 
-  const data = await prisma.todo.delete({
-    where: {
-      id: Number(id),
-    },
-  });
-  return NextResponse.json(data);
+    const data = await prisma.todo.delete({
+      where: {
+        id: parseInt(id, 10),
+      },
+    });
+
+    return NextResponse.json(data, { status: 200 });
+  } catch (error) {
+    console.error("Error deleting todo:", error);
+
+    return NextResponse.json(
+      { error: "Failed to delete the todo" },
+      { status: 500 }
+    );
+  }
 };
